Refactored APMBuildter tests.
